Restructure directories inside popup directory
